# specs/001-ai-consultation-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Car Service AI Assistant API
  version: 0.5.0
  description: >
    Полный операционный контракт, сгенерированный из runtime Express
    (`docs/api-route-inventory.json`). Схемы тел запросов — минимальные;
    источник истины по наличию path+method — инвентарь + `npm run openapi:check`.
servers:
  - url: http://localhost:3000/api
...
paths:
...
  "/contact":
    get:
      tags: [Contact]
      summary: Получить contact
      operationId: get_contact
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Contact]
      summary: Создать/выполнить contact
      operationId: post_contact
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/contact/{contactId}":
    patch:
      tags: [Contact]
      summary: "Обновить {contactId}"
      operationId: patch_contact_contactId
      parameters:
        - in: path
          name: contactId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/contact/{contactId}/convert-to-request":
    post:
      tags: [Contact]
      summary: Создать/выполнить convert-to-request
      operationId: post_contact_contactId_convert_to_request
      parameters:
        - in: path
          name: contactId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/content/site-items":
    get:
      tags: [Content]
      summary: Получить site-items
      operationId: get_content_site_items
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/content/site-settings":
    get:
      tags: [Content]
      summary: Получить site-settings
      operationId: get_content_site_settings
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/maintenance-alerts":
    get:
      tags: [Other]
      summary: Получить maintenance-alerts
      operationId: get_maintenance_alerts
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/service-records/{recordId}":
    delete:
      tags: [Other]
      summary: "Удалить {recordId}"
      operationId: delete_service_records_recordId
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      tags: [Other]
      summary: "Обновить {recordId}"
      operationId: patch_service_records_recordId
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/service-records/{recordId}/export.jpg":
    get:
      tags: [Other]
      summary: Получить export.jpg
      operationId: get_service_records_recordId_export_jpg
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/service-records/{recordId}/export.pdf":
    get:
      tags: [Other]
      summary: Получить export.pdf
      operationId: get_service_records_recordId_export_pdf
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/vehicles":
    get:
      tags: [Vehicles]
      summary: Получить vehicles
      operationId: get_vehicles
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Vehicles]
      summary: Создать/выполнить vehicles
      operationId: post_vehicles
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/vehicles/{vehicleId}/maintenance-plan":
    get:
      tags: [Vehicles]
      summary: Получить maintenance-plan
      operationId: get_vehicles_vehicleId_maintenance_plan
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  "/vehicles/{vehicleId}/service-history/export.pdf":
    get:
      tags: [Vehicles]
      summary: Получить export.pdf
      operationId: get_vehicles_vehicleId_service_history_export_pdf
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  "/vehicles/{vehicleId}/service-records":
    get:
      tags: [Vehicles]
      summary: Получить service-records
      operationId: get_vehicles_vehicleId_service_records
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      tags: [Vehicles]
      summary: Создать/выполнить service-records
      operationId: post_vehicles_vehicleId_service_records
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: OK
        '4XX':
          description: Problem Details (RFC 9457)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        error: { type: string }
        code: { type: string }
...